+++ app/src/main/AndroidManifest.xml
@@ -11,22 +11,65 @@
 
     <application
         android:allowBackup="true"
-        android:dataExtractionRules="@xml/data_extraction_rules"
-        android:fullBackupContent="@xml/backup_rules"
         android:icon="@mipmap/ic_launcher"
-        android:label="@string/app_name"
+        android:label="InvestikaTrial"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        tools:targetApi="31">
+        tools:targetApi="31"
+        android:theme="@style/Theme.InvestikaTrial">
 
         <!-- MainActivity definition -->
         <activity
-            android:name="com.akirachix.investika.ui.MainActivity"
+            android:name="com.akirachix.investikaTrial.ui.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
+        </activity>
+
+
+        <activity
+            android:name="com.akirachix.investikaTrial.ui.HighRiskPaths"
+            android:exported="false">
+
+        </activity>
+
+        <activity
+            android:name="com.akirachix.investikaTrial.ui.PathDecisionActivity"
+            android:exported="false">
+
+        </activity>
+
+        <activity
+            android:name="com.akirachix.investikaTrial.ui.BattleActivity"
+            android:exported="false">
+
+        </activity>
+
+        <activity
+            android:name="com.akirachix.investikaTrial.ui.BattleQuizActivity"
+            android:exported="false">
+
+        </activity>
+
+        <activity
+            android:name="com.akirachix.investikaTrial.ui.SlayDragonActivity"
+            android:exported="false">
+
+        </activity>
+
+
+        <activity
+            android:name="com.akirachix.investikaTrial.ui.DefeatedActivity"
+            android:exported="false">
+
+        </activity>
+
+        <activity
+            android:name="com.akirachix.investikaTrial.ui.ProfileActivity"
+            android:exported="false">
+
         </activity>
 
     </application>
