+++ app/src/main/java/com/akirachix/investikaTrial/ui/BattleActivity.kt
@@ -1,10 +1,9 @@
-package com.example.investika
+package com.akirachix.investikaTrial.ui
 
 import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
-import com.example.investika.databinding.ActivityBattleBinding
-import com.example.investika.databinding.ActivityMainBinding
+import com.akirachix.investikatrial.databinding.ActivityBattleBinding
 
 
 class BattleActivity : AppCompatActivity() {
