+++ app/src/main/java/com/akirachix/investikaTrial/ui/ProfileActivity.kt
@@ -1,0 +1,29 @@
+package com.akirachix.investikaTrial.ui
+
+import android.content.Intent
+import android.os.Bundle
+import androidx.activity.enableEdgeToEdge
+import androidx.appcompat.app.AppCompatActivity
+import androidx.core.view.ViewCompat
+import androidx.core.view.WindowInsetsCompat
+import com.akirachix.investikatrial.databinding.ActivityMainBinding
+import com.akirachix.investikatrial.databinding.ActivityProfileBinding
+
+class ProfileActivity : AppCompatActivity() {
+
+    private lateinit var binding: ActivityProfileBinding
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        binding = ActivityProfileBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+
+        binding.button.setOnClickListener {
+            val intent = Intent(this, HighRiskPaths::class.java)
+            startActivity(intent)
+        }
+    }
+
+
+}
