--- app/src/main/java/com/example/investika/MainActivity.kt
@@ -1,23 +1,0 @@
-package com.example.investika
-
-import android.content.Intent
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import com.example.investika.databinding.ActivityMainBinding
-
-
-class MainActivity : AppCompatActivity() {
-    private lateinit var binding: ActivityMainBinding
-
-        override fun onCreate(savedInstanceState: Bundle?) {
-            super.onCreate(savedInstanceState)
-
-        binding = ActivityMainBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-        binding.button.setOnClickListener {
-            val intent = Intent(this, HighRiskPaths::class.java)
-            startActivity(intent)
-        }
-    }
-}
