--- app/src/main/res/drawable/boarder_radius.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="rectangle">
-    <corners android:radius="10dp" />
-    <solid android:color="@android:color/white" /> <!-- Transparent background -->
-    <stroke android:width="2dp" android:color="#FFFFFF" /> <!-- Optional border color -->
-</shape>
