+++ app/src/main/res/drawable/corners.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+    <!-- Set the corner radius here -->
+    <corners android:radius="16dp" />
+    <!-- Optional: Set solid color or gradient -->
+    <solid android:color="#FFFFFF" /> <!-- Button background color -->
+    <!-- Optional: Set border/stroke -->
+    <stroke
+        android:width="2dp"
+        android:color="#FFFFFF" /> <!-- Border color and width -->
+</shape>
+
+
+
+
