+++ app/src/main/res/layout/activity_battle.xml
@@ -6,18 +6,16 @@
     android:layout_height="match_parent"
     android:padding="16dp"
     android:gravity="center"
-    android:background="@drawable/riverbg"
-    tools:context=".MainActivity">
-
+    android:background="@drawable/riverbg">
 
 
     <TextView
         android:id="@+id/textView"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginTop="36dp"
-        android:text="The battle begins!!"
+        android:layout_marginTop="30dp"
         android:fontFamily="@font/nunito_bold"
+        android:text="The battle begins!!"
         android:textColor="#FFFFFF"
         android:textFontWeight="@integer/material_motion_duration_medium_2"
         android:textSize="24sp"
@@ -26,17 +24,17 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
+
     <ImageView
         android:id="@+id/imageView"
         android:layout_width="298dp"
         android:layout_height="263dp"
-        android:layout_marginTop="24dp"
+        android:layout_marginTop="40dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.493"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/textView"
         app:srcCompat="@drawable/soldiers" />
-
 
     <TextView
         android:id="@+id/textView2"
