+++ app/src/main/res/layout/activity_battlequiz.xml
@@ -28,8 +28,9 @@
         android:layout_width="350dp"
         android:layout_height="65dp"
         android:layout_marginBottom="70dp"
-        android:background="@drawable/boarder_radius"
         android:backgroundTint="#FFFFFF"
+        android:background="@drawable/corners"
+        android:fontFamily="@font/nunito_bold"
         android:text="Put money in one place"
         android:textColor="@color/black"
         android:textSize="20sp"
@@ -42,10 +43,11 @@
         android:layout_width="350dp"
         android:layout_height="65dp"
         android:layout_marginBottom="60dp"
-        android:background="@drawable/boarder_radius"
         android:backgroundTint="#FFFFFF"
         android:text="Borrow money"
+        android:fontFamily="@font/nunito_bold"
         android:textColor="@color/black"
+        android:background="@drawable/corners"
         android:textSize="20sp"
         app:layout_constraintBottom_toTopOf="@+id/button15"
         app:layout_constraintEnd_toEndOf="parent"
@@ -57,8 +59,9 @@
         android:layout_width="350dp"
         android:layout_height="65dp"
         android:layout_marginBottom="60dp"
-        android:background="@drawable/boarder_radius"
         android:backgroundTint="#FFFFFF"
+        android:background="@drawable/corners"
+        android:fontFamily="@font/nunito_bold"
         android:text="Keep cash at home"
         android:textColor="@color/black"
         android:textSize="20sp"
@@ -72,7 +75,8 @@
         android:layout_width="350dp"
         android:layout_height="65dp"
         android:layout_marginBottom="60dp"
-        android:background="@drawable/boarder_radius"
+        android:background="@drawable/corners"
+        android:fontFamily="@font/nunito_bold"
         android:backgroundTint="#FFFFFF"
         android:text="Spread money across investments"
         android:textColor="@color/black"
