+++ app/src/main/res/layout/activity_defeated.xml
@@ -6,8 +6,7 @@
     android:layout_height="match_parent"
     android:padding="16dp"
     android:gravity="center"
-    android:background="@drawable/riverbg"
-    tools:context=".MainActivity">
+    android:background="@drawable/riverbg">
 
 
 
@@ -50,18 +49,18 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
-
     <Button
-        android:id="@+id/button"
-        android:layout_width="216dp"
-        android:layout_height="62dp"
-        android:layout_marginStart="80dp"
-        android:layout_marginBottom="40dp"
-        android:backgroundTint="#FFC300"
+        android:id="@+id/btnProceed"
+        android:layout_width="200dp"
+        android:layout_height="wrap_content"
+        android:layout_marginBottom="44dp"
+        android:backgroundTint="#FFC107"
+        android:fontFamily="@font/nunito_bold"
         android:text="Continue"
         android:textColor="#003566"
-        android:textSize="20sp"
+        android:textSize="20dp"
         app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.497"
         app:layout_constraintStart_toStartOf="parent" />
-
 </androidx.constraintlayout.widget.ConstraintLayout>
