+++ app/src/main/res/layout/activity_path_decision.xml
@@ -28,10 +28,12 @@
         android:layout_width="350dp"
         android:layout_height="65dp"
         android:layout_marginBottom="60dp"
-        android:background="@drawable/boarder_radius"
+        android:background="@drawable/corners"
         android:backgroundTint="#FFFFFF"
         android:text="I'm confident in my choice"
+        android:fontFamily="@font/nunito_bold"
         android:textColor="@color/black"
+        android:textColorLink="#FFFFFF"
         android:textSize="20sp"
         app:layout_constraintBottom_toTopOf="@+id/button12"
         app:layout_constraintEnd_toEndOf="@+id/button12"
@@ -43,9 +45,10 @@
         android:layout_width="350dp"
         android:layout_height="65dp"
         android:layout_marginBottom="60dp"
-        android:background="@drawable/boarder_radius"
         android:backgroundTint="#FFFFFF"
+        android:fontFamily="@font/nunito_bold"
         android:text="I want quick gains"
+        android:background="@drawable/corners"
         android:textColor="@color/black"
         android:textSize="20sp"
         app:layout_constraintBottom_toTopOf="@+id/button15"
@@ -58,9 +61,10 @@
         android:layout_width="350dp"
         android:layout_height="65dp"
         android:layout_marginBottom="88dp"
-        android:background="@drawable/boarder_radius"
         android:backgroundTint="#FFFFFF"
         android:text="Aware of negative outcomes"
+        android:fontFamily="@font/nunito_bold"
+        android:background="@drawable/corners"
         android:textColor="@color/black"
         android:textSize="20sp"
         app:layout_constraintBottom_toBottomOf="parent"
@@ -73,9 +77,10 @@
         android:layout_width="350dp"
         android:layout_height="65dp"
         android:layout_marginBottom="60dp"
-        android:background="@drawable/boarder_radius"
         android:backgroundTint="#FFFFFF"
         android:text="High Risk, high reward"
+        android:fontFamily="@font/nunito_bold"
+        android:background="@drawable/corners"
         android:textColor="@color/black"
         android:textSize="20sp"
         app:layout_constraintBottom_toTopOf="@+id/button16"
