+++ app/src/main/res/layout/activity_slaydragon.xml
@@ -6,8 +6,7 @@
     android:layout_height="match_parent"
     android:padding="16dp"
     android:gravity="center"
-    android:background="@drawable/riverbg"
-    tools:context=".MainActivity">
+    android:background="@drawable/riverbg">
 
 
 
@@ -34,7 +33,7 @@
         app:layout_constraintHorizontal_bias="0.493"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/textView"
-        app:srcCompat="@drawable/coins" />
+        app:srcCompat="@drawable/giftcoins" />
 
     <TextView
         android:id="@+id/textView2"
